Server/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
